move to inc directory

+++ compiler/inc/error_handler.h
@@ -1,7 +1,6 @@
 #ifndef ERROR_HANDLER_H
 #define ERROR_HANDLER_H
 
-#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
@@ -11,7 +10,7 @@
 
 //  ===-------------------------------------------=== FUNC DEC
 //  ***---------------------*** VOID FUNCS
-void err_ex_p(const char *_err_msg, const char *__FILE, uint32_t __LINE);
-void todo_err(const char *_msg, const char *__FILE, uint32_t __LINE);
+void err_ex_p(const char *_err_msg, const char *__FILE, size_t __LINE);
+void todo_err(const char *_msg, const char *__FILE, size_t __LINE);
 
 #endif
